Hexadec nbrs negative nbrs fixed

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: skunert <skunert@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/03 11:34:29 by skunert           #+#    #+#             */
-/*   Updated: 2023/04/05 10:25:38 by skunert          ###   ########.fr       */
+/*   Updated: 2023/04/05 10:32:21 by skunert          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void	ft_putchar(char c, int *ret_len);
 void	ft_putstr(char *s, int *ret_len);
 void	ft_putpercent(int *ret_len);
 void	ft_putnbr(long long n, int *ret_len);
-void	ft_putnbr_hexa_lower(long n, int *ret_len);
-void	ft_putnbr_hexa_upper(long n, int *ret_len);
+void	ft_putnbr_hexa_lower(long long n, int *ret_len);
+void	ft_putnbr_hexa_upper(long long n, int *ret_len);
 
 #endif
